fixed type for longitoa

+++ ft_longitoa_base.c
@@ -6,13 +6,13 @@
 /*   By: dpiedra <dpiedra@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/04 11:42:40 by dpiedra           #+#    #+#             */
-/*   Updated: 2020/02/06 11:52:03 by dpiedra          ###   ########.fr       */
+/*   Updated: 2020/02/06 17:02:20 by dpiedra          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "printf.h"
 
-int		ft_longcount_space(int base, int i)
+unsigned long long	ft_longcount_space(int base, int i)
 {
 	if (i == 0)
 		return (1);
@@ -20,7 +20,7 @@ int		ft_longcount_space(int base, int i)
 		return (base * ft_count_space(base, i - 1));
 }
 
-char	*ft_longitoa_base(unsigned long long nbr, int base)
+char				*ft_longitoa_base(unsigned long long nbr, int base)
 {
 	int		i;
 	char	*str;
